pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aurora-web"
version = "0.1.0"
description = "Python web interface for LED matrix control"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
include = ["aurora_web*"]

[tool.pytest.ini_options]
testpaths = ["aurora_web/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
